requirements.txt
Flask>=2.0
python-dotenv>=1.0
psycopg2-binary>=2.9
waitress>=2.1
boto3>=1.26
